% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{insight.cast.odin-package}
\alias{insight.cast.odin}
\alias{insight.cast.odin-package}
\title{insight.cast.odin: use odin2 models with insight.cast}
\description{
The main functions are \code{\link[=odin_sir]{odin_sir()}} for the ready-to-use model and \code{\link[=ODIN2]{ODIN2()}}
for a model written by the user.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ACCIDDA/insight.cast.odin}
  \item \url{https://accidda.github.io/insight.cast.odin/}
  \item Report bugs at \url{https://github.com/ACCIDDA/insight.cast.odin/issues}
}

}
\author{
\strong{Maintainer}: Cyril Geismar \email{cgeisma1@jh.edu}

Other contributors:
\itemize{
  \item Insight Net [copyright holder, funder]
}

}
\keyword{internal}
